requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==1.3.7
pydantic==2.5.0
beautifulsoup4==4.12.2
pytz==2023.3
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
ruff==0.1.7
